Refactor l'interface d'authentification : suppression des composants Suspense et mise à jour du pied de page
