+++ cms/src/admin/app.example.tsx
@@ -32,6 +32,6 @@ export default {
     ],
   },
   bootstrap(app: StrapiApp) {
-    console.log(app);
+    // console.log(app);
   },
 };
