+++ frontend/app/[locale]/(auth)/sign-in/page.tsx
@@ -1,8 +1,7 @@
+import { AuthFooter } from "@/components/auth/auth-footer";
 import { AuthHeader } from "@/components/auth/auth-header";
 import SignIn from "@/components/auth/signin";
 import fetchContentType, { StrapiData } from "@/lib/strapi/fetchContentType";
-import Link from "next/link";
-// import { redirect } from "next/navigation";
 
 type SignInPageProps = Promise<{
   locale: string;
@@ -24,9 +23,7 @@ export default async function SignInPage(props: { params: SignInPageProps }) {
     supportedLanguages.forEach((language: StrapiData) => {
       locales.push(language.name as string);
     });
-    console.log(supportedLanguages);
   }
-  console.log(locales);
 
   const signInPageContent = (await fetchContentType(
     "signins",
@@ -44,19 +41,7 @@ export default async function SignInPage(props: { params: SignInPageProps }) {
       <main className="flex grow items-center justify-center px-4">
         <SignIn data={signInPageContent} />
       </main>
-      <footer className="w-full bg-gray-200 p-4 dark:bg-gray-800">
-        <div className="container mx-auto text-center text-sm text-gray-600 dark:text-gray-400">
-          <p>&copy; 2023 Your Company Name. All rights reserved.</p>
-          <div className="mt-2">
-            <Link href="/privacy" className="mr-4 hover:underline">
-              Privacy Policy
-            </Link>
-            <Link href="/terms" className="hover:underline">
-              Terms of Service
-            </Link>
-          </div>
-        </div>
-      </footer>
+      <AuthFooter />
     </div>
   );
 }
