+++ frontend/app/[locale]/layout.tsx
@@ -1,4 +1,4 @@
-import { ReactNode, Suspense } from "react";
+import { ReactNode } from "react";
 
 interface LayoutProps {
   children: ReactNode;
@@ -6,18 +6,6 @@ interface LayoutProps {
 
 export default async function Layout({ children }: LayoutProps) {
   return (
-    <Suspense fallback={<SuspenseFallback />}>
-      <main className="flex size-full flex-col bg-background">{children}</main>
-    </Suspense>
-  );
-}
-
-function SuspenseFallback() {
-  return (
-    <div className="flex h-screen w-full items-center justify-center">
-      <div className="flex flex-col items-center space-y-4">
-        <div className="size-32 animate-spin rounded-full border-4 border-gray-300 border-t-gray-900" />
-      </div>
-    </div>
+    <main className="flex size-full flex-col bg-background">{children}</main>
   );
 }
