+++ frontend/components/auth/auth-footer.tsx
@@ -5,7 +5,7 @@ export const AuthFooter = () => {
   return (
     <footer className="w-full bg-gray-200 p-4 dark:bg-gray-800">
       <div className="container mx-auto text-center text-sm text-gray-600 dark:text-gray-400">
-        <p>&copy; 2023 Your Company Name. All rights reserved.</p>
+        <p>&copy; 2024 Gouale e.i. All rights reserved.</p>
         <div className="mt-2">
           <Link href="/privacy" className="mr-4 hover:underline">
             Privacy Policy
